+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Enhancements
 
 * `SmartString` now uses `LazyCompact` instead of `Compact` to minimize allocations.
 * Added `pop` for strings.
+* Added `ImmutableString::ptr_eq` to test if two strings point to the same allocation.
 
 ### `Scope` API